refactor(songs): move new components to their own files
