+++ lib/songs/player/song_player.dart
@@ -3,90 +3,12 @@ import 'dart:io';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_soloud/flutter_soloud.dart';
-// ignore: implementation_imports
-import 'package:flutter_soloud/src/filters/pitchshift_filter.dart';
-// ignore: implementation_imports
-import 'package:flutter_soloud/src/filters/equalizer_filter.dart';
-import 'package:musbx/songs/musbx_api/demixer_api.dart';
+import 'package:musbx/songs/demixer/demixer.dart';
+import 'package:musbx/songs/equalizer/equalizer.dart';
 import 'package:musbx/songs/player/playable.dart';
 import 'package:musbx/songs/player/song.dart';
+import 'package:musbx/songs/slowdowner/slowdowner.dart';
 import 'package:musbx/widgets/widgets.dart';
-
-class Stem {
-  /// The default [volume]
-  static const double defaultVolume = 0.5;
-
-  static final SoLoud _soloud = SoLoud.instance;
-
-  /// A demixed stem for a song. Can be played back in sync with other stems.
-  ///
-  /// There should (usually) only ever be one stem of each [type].
-  Stem(this.type, this.player);
-
-  /// The player that this is a part of.
-  final SongPlayer player;
-
-  /// The type of stem.
-  final StemType type;
-
-  /// The source of the stem of the [player]'s [Playable] with the same [type] as this, if it is a [DemixedAudio].
-  AudioSource? get source {
-    if (player.playable is! DemixedAudio) return null;
-
-    return (player.playable as DemixedAudio).sources[type];
-  }
-
-  /// The handle of the stem of the [player]'s [Playable] with the same [type] as this, if it is a [DemixedAudio].
-  SoundHandle? get handle {
-    if (player.playable is! DemixedAudio) return null;
-
-    return (player.playable as DemixedAudio).handles?[type];
-  }
-
-  /// Whether this stem is enabled and should be played.
-  ///
-  /// If this is `false` the audio will be muted, regardless of the value of [volume].
-  bool get enabled => enabledNotifier.value;
-  set enabled(bool value) => enabledNotifier.value = value;
-  late final ValueNotifier<bool> enabledNotifier = ValueNotifier(true)
-    ..addListener(_updateEnabled);
-
-  void _updateEnabled() {
-    final SoundHandle? handle = this.handle;
-    if (handle == null) return;
-
-    if (enabled) {
-      _soloud.setVolume(handle, volume);
-    } else {
-      _soloud.setVolume(handle, 0.0);
-    }
-  }
-
-  /// The volume this stem is played at. The value is clamped between 0 and 1.
-  double get volume => volumeNotifier.value;
-  set volume(double value) => volumeNotifier.value = value.clamp(0, 1);
-  late final ValueNotifier<double> volumeNotifier = ValueNotifier(defaultVolume)
-    ..addListener(_updateVolume);
-
-  void _updateVolume() {
-    final SoundHandle? handle = this.handle;
-    if (handle == null) return;
-
-    if (enabled) {
-      _soloud.setVolume(handle, volume);
-    }
-  }
-}
-
-class StemsNotifier extends ValueNotifier<List<Stem>> {
-  /// Notifies listeners whenever [enabled] or [volume] of any of the stems provided in [value] changes.
-  StemsNotifier(super.value) {
-    for (Stem stem in value) {
-      stem.enabledNotifier.addListener(notifyListeners);
-      stem.volumeNotifier.addListener(notifyListeners);
-    }
-  }
-}
 
 abstract class SongPlayerComponent {
   SongPlayerComponent(this.player);
@@ -117,262 +39,6 @@ abstract class SongPlayerComponent {
   @mustCallSuper
   Map<String, dynamic> saveSettingsToJson() {
     return {};
-  }
-}
-
-class SongDemixer extends SongPlayerComponent {
-  /// A component of the [SongPlayer] that is used to separate a song into stems and change the volume of those individually.
-  SongDemixer(
-    super.player,
-  );
-
-  /// The stems that this song has been separated into.
-  List<Stem> get stems => stemsNotifier.value;
-  late final StemsNotifier stemsNotifier = StemsNotifier(List.unmodifiable([
-    Stem(StemType.drums, player),
-    Stem(StemType.piano, player),
-    Stem(StemType.guitar, player),
-    Stem(StemType.bass, player),
-    Stem(StemType.vocals, player),
-    Stem(StemType.other, player),
-  ]));
-
-  /// Load settings from a [json] map.
-  ///
-  /// [json] can contain the following stems:
-  ///  - `drums`
-  ///  - `piano`
-  ///  - `guitar`
-  ///  - `bass`
-  ///  - `vocals`
-  ///  - `other`
-  ///
-  /// Each stem can contain the following key-value pairs:
-  ///  - `enabled` [bool] Whether this stem is enabled and should be played.
-  ///  - `volume` [double] The volume this stem is played back at. Must be between 0 and 1.
-  @override
-  void loadSettingsFromJson(Map<String, dynamic> json) {
-    super.loadSettingsFromJson(json);
-
-    for (Stem stem in stems) {
-      Map<String, dynamic>? stemData =
-          tryCast<Map<String, dynamic>>(json[stem.type.name]);
-
-      bool? enabled = tryCast<bool>(stemData?["enabled"]);
-      stem.enabled = enabled ?? true;
-
-      double? volume = tryCast<double>(stemData?["volume"]);
-      stem.volume = volume ?? 0.5;
-    }
-  }
-
-  /// Save settings for a song to a json map.
-  ///
-  /// Saves the following stems:
-  ///  - `drums`
-  ///  - `piano`
-  ///  - `guitar`
-  ///  - `bass`
-  ///  - `vocals`
-  ///  - `other`
-  ///
-  /// Each stem contains the following key-value pairs:
-  ///  - `enabled` [bool] Whether this stem is enabled and should be played.
-  ///  - `volume` [double] The volume this stem is played back at. Must be between 0 and 1.
-  @override
-  Map<String, dynamic> saveSettingsToJson() {
-    return {
-      ...super.saveSettingsToJson(),
-      for (Stem stem in stems)
-        stem.type.name: {
-          "enabled": stem.enabled,
-          "volume": stem.volume,
-        }
-    };
-  }
-}
-
-class SongSlowdowner extends SongPlayerComponent {
-  SongSlowdowner(super.player);
-
-  /// Modify the pitch filter for the current song.
-  ///
-  /// Note that this cannot be used to activate or deactivate a filter, since
-  /// that has to be done before the song is loaded and [player.handle] thus
-  /// isn't available at that time.
-  void _modifyPitchFilter(
-      void Function(PitchShiftSingle filter, {SoundHandle? handle}) modify) {
-    player.playable.filters(handle: player.handle).pitchShift.modify(modify);
-  }
-
-  @override
-  void initialize() {
-    player.playable.filters().pitchShift.activate();
-  }
-
-  @override
-  void dispose() {
-    player.playable.filters().pitchShift.deactivate();
-  }
-
-  /// How much the pitch will be shifted, in semitones.
-  double get pitch => pitchNotifier.value;
-  set pitch(double value) => pitchNotifier.value = value;
-  late final ValueNotifier<double> pitchNotifier = ValueNotifier(0)
-    ..addListener(_updatePitch);
-
-  void _updatePitch() {
-    _modifyPitchFilter((filter, {SoundHandle? handle}) {
-      filter.semitones(soundHandle: handle).value = pitch;
-    });
-  }
-
-  /// The playback speed.
-  double get speed => speedNotifier.value;
-  set speed(double value) => speedNotifier.value = value;
-  late final ValueNotifier<double> speedNotifier = ValueNotifier(1)
-    ..addListener(_updateSpeed);
-
-  void _updateSpeed() {
-    SoLoud.instance.setRelativePlaySpeed(player.handle, speed);
-    _modifyPitchFilter((filter, {SoundHandle? handle}) {
-      filter.shift(soundHandle: handle).value = 1 / speed;
-    });
-  }
-
-  /// Load settings from a [json] map.
-  ///
-  /// [json] can contain the following key-value pairs:
-  ///  - `pitch` [double] How much the pitch will be shifted, in semitones.
-  ///  - `speed` [double] The playback speed of the audio, as a fraction.
-  @override
-  void loadSettingsFromJson(Map<String, dynamic> json) {
-    super.loadSettingsFromJson(json);
-
-    pitch = tryCast<double>(json["pitch"])?.clamp(-12, 12) ?? 0.0;
-    speed = tryCast<double>(json["speed"])?.clamp(0.5, 2) ?? 1.0;
-  }
-
-  /// Save settings for a song to a json map.
-  ///
-  /// Saves the following key-value pairs:
-  ///  - `pitch` [double] How much the pitch will be shifted, in semitones.
-  ///  - `speed` [double] The playback speed of the audio, as a fraction.
-  @override
-  Map<String, dynamic> saveSettingsToJson() {
-    return {
-      ...super.saveSettingsToJson(),
-      "pitch": pitch,
-      "speed": speed,
-    };
-  }
-}
-
-class EqualizerBand {
-  /// The minimum value for the [gain].
-  static const double minGain = 0.0;
-
-  /// The maximum value for the [gain].
-  static const double maxGain = 4.0;
-
-  /// The gain for this band. Must be between [minGain] and [maxGain] (since that is the range SoLoud allows).
-  double get gain => gainNotifier.value;
-  set gain(double value) => gainNotifier.value = value.clamp(minGain, maxGain);
-  late final ValueNotifier<double> gainNotifier = ValueNotifier(1.0);
-}
-
-class EqualizerBandsNotifier extends ValueNotifier<List<EqualizerBand>> {
-  /// Notifies listeners whenever [gain] of any of the bands provided in [value] changes.
-  EqualizerBandsNotifier(super.value) {
-    for (EqualizerBand band in value) {
-      band.gainNotifier.addListener(notifyListeners);
-    }
-  }
-}
-
-class SongEqualizer extends SongPlayerComponent {
-  static const double defaultGain = 1.0;
-
-  /// TODO: Simply activating this causes lots of artifacts at the moment, not sure why
-  SongEqualizer(super.player);
-
-  /// Modify the equalizer filter for the current song.
-  ///
-  /// Note that this cannot be used to activate or deactivate the filter, since
-  /// that has to be done before the song is loaded and [player.handle] thus
-  /// isn't available at that time.
-  void _modifyEqualizerFilter(
-      void Function(EqualizerSingle filter, {SoundHandle? handle}) modify) {
-    player.playable.filters(handle: player.handle).equalizer.modify(modify);
-  }
-
-  @override
-  void initialize() {
-    // TODO: Activate the equalizer once we find out what is causing the noise
-    // player.playable.filters().pitchShift.activate();
-  }
-
-  @override
-  void dispose() {
-    player.playable.filters().pitchShift.deactivate();
-  }
-
-  /// The frequency bands of the equalizer.
-  List<EqualizerBand> get bands => bandsNotifier.value;
-  late final EqualizerBandsNotifier bandsNotifier = EqualizerBandsNotifier(
-      List.unmodifiable(List.generate(8, (index) => EqualizerBand())))
-    ..addListener(_updateBands);
-
-  void _updateBands() {
-    _modifyEqualizerFilter((filter, {SoundHandle? handle}) {
-      for (int i = 0; i < bands.length; i++) {
-        [
-          filter.band1,
-          filter.band2,
-          filter.band3,
-          filter.band4,
-          filter.band5,
-          filter.band6,
-          filter.band7,
-          filter.band8,
-        ][i](soundHandle: handle)
-            .value = bands[i].gain;
-      }
-    });
-  }
-
-  /// Reset the gain on all [bands].
-  void resetGain() {
-    for (var band in bands) {
-      band.gain = defaultGain;
-    }
-  }
-
-  /// Load settings from a [json] map.
-  ///
-  /// [json] can contain the following key-value pairs (beyond `enabled`):
-  ///  - `gain` [Map<String, double>] The gain for the frequency bands, with the key being the index of the band (usually 0-4) and the value being the gain.
-  @override
-  void loadSettingsFromJson(Map<String, dynamic> json) async {
-    super.loadSettingsFromJson(json);
-
-    final Map? gains = tryCast<Map>(json["gain"]);
-    for (var i = 0; i < bands.length; i++) {
-      final double gain = tryCast<double>(gains?["$i"]) ?? defaultGain;
-      bands[i].gain = gain;
-    }
-  }
-
-  /// Save settings for a song to a json map.
-  ///
-  /// Saves the following key-value pairs (beyond `enabled`):
-  ///  - `gain` [Map<String, double>] The gain for the frequency bands, with the key being the index of the band (usually 0-4) and the value being the gain.
-  @override
-  Map<String, dynamic> saveSettingsToJson() {
-    return {
-      ...super.saveSettingsToJson(),
-      "gain": bands.asMap().map((index, band) => MapEntry("$index", band.gain)),
-    };
   }
 }
 
@@ -489,13 +155,13 @@ class SongPlayer {
   ];
 
   /// Component for isolating or muting specific instruments in the song.
-  late final SongDemixer demixer = SongDemixer(this);
+  late final DemixerComponent demixer = DemixerComponent(this);
 
   /// Component for changing the pitch and speed of the song.
-  late final SongSlowdowner slowdowner = SongSlowdowner(this);
+  late final SlowdownerComponent slowdowner = SlowdownerComponent(this);
 
   /// Component for adjusting the gain for different frequency bands of the song.
-  late final SongEqualizer equalizer = SongEqualizer(this);
+  late final EqualizerComponent equalizer = EqualizerComponent(this);
 
   /// Load song preferences from a [json] map.
   void loadPreferences(Map<String, dynamic> json) {
